Introduce server intermediary
I'd like to be able to access my password db through a website.
The primary goal for this change is to create a new HTML client
for passphrase. It also opens up the opportunity of accessing a
LAN passphrase server through, say my phone.

Since there are no secrets in the DB and I'm the only client, the
messages are unsecured.